Proxy secure notifications/interactions through Player.
Interactions are completed by publishing a reply to the notification feed and
sending data to its associated working directory. Both the feed and the working
directory are secured with the standard Taverna Server credentials so we need to
proxy all this communication through the Rails app (Player) so that we can use
the correct credentials.