Move setting run status to "finished" until it *really* is!
We have a different idea as to when a run is finished than Taverna Server. For
us it's not done until the results are saved and the post-run tasks are
complete.

Moving the setting of the finished status to the very end of all this ensures
that we avoid a race condition in the UI.