[TAV-340] Run progress auto-refresh.
Poll for run updates at a rate of half that of the Taverna Server poll rate.
Update any run info, show any interactions and display outputs when they arrive.

Tests included for interactions and new interactions in the runs controller and
the displayed flag in the interaction model.

Player now need jquery so uncomment it from the dependencies.