[TAV-426] Fix interaction race condition.
When getting interactions from Taverna Server there was a race condition whereby
the interaction was being saved in the database before it was fully setup with
all the required data. If the web front end happened to get this interaction
before this data was added then it would be shown as a blank interaction.