Ensure the proxied models use a rooted (::) namespace.
This makes sure that there are no clashes with similar names. Without this it
is possible that proxy names could end up either in, or clashing with names in,
the TavernaPlayer namespace. This can probably only happen with Workflow <->
TavernaPlayer::Workflow at the moment, but this is safe now in the future.